update get cpu time

+++ data_structures/heap/skew_heap_test.c
@@ -46,7 +46,7 @@ static void testSkewHeap() {
     }
     printSkewHeap(skewHeap);
     end = getTime();
-    printf("duration = %lfs\n", (double) calcDuration(start, end));
+    printf("duration = %lfs\n", calcDuration(start, end));
 
     skewHeapDestroy(skewHeap);
 }
