+++ data_structures/stack_queue/queue_array/queue_test.c
@@ -23,7 +23,7 @@ static void testQueue() {
     }
     end = getTime();
     printf("size of queue: %d\n", queueSize(queue));
-    printf("duration = %lfs\n", (double) calcDuration(start, end));
+    printf("duration = %lfs\n", calcDuration(start, end));
 
     if (!areArraysSame(tmpArray, array, ARRAY_SIZE)) {
         fprintf(stderr, "Array aren't same.\n");
